Arrow Changes
An Arrow now gets created, but not in the right place. It also gets deleted now before new ones are created. I am working on creating a new Arrow Object to use rather than an image. Hopefully this works better than the image in the direction it faces when created.